refactor: manager 관련 파일 계층화하여 디렉토리 구조 변경
